Add files via upload
This program allows users to look up, add, change, or delete an employee with the "choice" file. "Employee" is the file where all the employee's information is stored. "ShiftSupervisor" file determines if that employee will be receiving a bonus and output their salary. "ProdWorker" is where you input the employee's information to store it in the "Employee" file.